fixed re-installing an updated mod would create a "duplicate" mod entry
When updating a mod and selecting "Update for all profiles" we maintain
the id and installation folder of the older mod in order to... well,
update the mod for all the profiles. Problem with that is if the user
then re-installs the updated mod for whatever reason, Vortex would fail
to realize that the mod is already installed - thus creating a duplicate
mod folder in staging.