fixed randomization of crash dump implementation
we have two implementations of generating crash dumps when vortex crashes, our own and the one included in electron.
We don't fully trust the electron one because it's been bugged and not doing anything in the past but it would be great if we could drop our own implementation to reduce the number of dependencies.
So atm we chose 50/50 which one to use in the hope of getting some data on whether they both work the same.