fixed: freeze/crash if category data from API is corrupted
specifically if the child-parent relationship between categories form a loop.

This is a two part fix:
a) The core app/bundled collections should now deal gracefully with invalid data
b) upon importing data from the api we break up any loops

b is there so that extensions trying to use the data don't have to suffer the same madness we do.