disabled "task" list
due to a bug in electron it doesn't actually get loaded most of the time but apparently under some circumstances it does.
Besides that the list doesn't filter out disabled tools but considering it doesn't work correctly anyway it'd be a bit pointless to try and fix that at this time.

fixes #13594