fixed: unhandled error if discovery update fails
This occurred when identifying the store for a manually selected game folder and when checking if a total conversion replaces the base game.
In these cases we didn't catch exceptions - not that these exceptions are supposed to happen, the only known cases come from users using Vortex through wine.

fixes #13770