make extra sure we never group collections
"group" in this context means combining different versions of a mod on the same line with a dropdown for the version number.
We currently don't allow multiple versions of a collection to be installed at the same time and the api doesn't support collections containing different variants so the only way collections would ever be grouped is caused by invalid data.