fixed inability to switch profiles when using symlink activator
The pre-existence of a userGate instance (which blocks deployment
events until a notification is interacted with) was blocking
game/profile changes when such a notification existed, the entirety of
the UX surrounding this piece of functionality is redundant as the
Windows elevation dialog would basically ask the same thing.

Elevation notification has been removed - Vortex will now raise the
WinAPI elevation dialog by default.