fixed inability to update mods that are tagged as compatibleDownloads
FO4 for example, provides the ability to download FOLON mods.

The update check mechanism was attempting to check for an update
using the FO4 domain name regardless of the domain name used when
the mod was initially downloaded.

This would obviously result in Vortex installing a different mod as an
update..