fixed mods update error handling
Not sure how it's possible to attempt to update mods for a game which is
not stored in Vortex's state; but the error handling was flawed anyway,
and it was possible to attempt to start a download without knowing which
game we're downloading for!

I guess it's possible for the game to become unmanaged during update
attempts (user removes extension or choosing to unmanage the game) but
this is a pretty weird one either way.

Also ensured we have a proper fallback in place which will open the website
url for that mod instead of crashing.

fixes nexus-mods/vortex#16251