fixed raw request error handling
Unsuccessful requests (statusCode !== 200) would resolve without
reporting an error - this would cause components which are using this
utility function to assume that the request was successful and continue
running its logic.

Most prominently - this could potentially replace the masterlist for
gamebryo games with rubbish data, making it impossible to sort plugins.

https://discord.com/channels/215154001799413770/408252140533055499/1273420884975292569