UV thread pool size is now defined as soon as the app is run
Hyper threading assures that 1 physical core is seen as 2 logical cores,
and we therefore now use 2 instead of 1.5 when calculating the size of
the thread pool. This should speed up file operations at least slightly.

Additionally, we now set the size as soon as the app starts to ensure
that NodeJS picks up the change from the get-go