added error details sanitization to prevent crashpad issues
- only keep printable ASCII characters
- truncated error meessage just in case
- pre-computing hex conversion to avoid potential issues during error
  creation

Please note: the actual issue we want to fix is the LOOT process failing
to start, but the error reports are incomplete due to the crash pad and
we therefore cannot debug/check why the process dies.

We should be able to get the actual error message in future releases.

fixes nexus-mods/vortex#18251
fixes nexus-mods/vortex#18250