adding backwards compatibility for old translation separators
with the new ':' and '.' separators it's impossible for us to ascertain
whether we're translating a string literal or are using a namespace key.

this causes i18n to truncate string literals whenever it encounters a
namespace/key separator.

given that 99% of our codebase is using string literals, we need to
ensure that we're backwards compatible.

this fixes current translations app-wide
- fixes string truncations for all dialogs and notifications
- fixes broken translation extensions
- fixes crash/bug report truncations
- fixes broken translation creation extension

closes nexus-mods/vortex#18768